% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoolingPeriod.R
\name{CoolingPeriod}
\alias{CoolingPeriod}
\title{Cooling Period}
\usage{
CoolingPeriod(x, res = 1, model)
}
\arguments{
\item{x}{The soil depth in cm.}

\item{res}{The temporal resolution (in min) at which to calculate the temperatures. Default is one minute.}

\item{model}{The model object (a list) from the SheFire function containing the equations and data that comprise the model.}
}
\value{
The function returns the temperatures at the specified time interval (res) for the soil depth (x) from the point that the soil depth begins to cool through the end of the model time range.  It also prints a text string summarizing some information about the cooling.
}
\description{
Isolates the period of time when the soil depth of interest is cooling and returns the temperature values for that period.
}
\details{
Example data from US Forest Service Fire Behavior Assessment Team: https://www.fs.fed.us/adaptivemanagement/reports/fbat/2019_FBATReport_WalkerFire_10112019_Final.pdf
}
\examples{
inputFile <- read.csv(file.path(system.file("external", package="SheFire"), "WalkerPlot4NE.csv")) #read in example data
model_example <- SheFire(input = inputFile )  #create model with all default settings

CoolingDown <- CoolingPeriod(x = 5, res = 1, model = model_example)
}
\seealso{
\code{\link{SheFire}}, \code{\link{HeatingPeriod}}
}
\author{
MaryKBrady
}
\keyword{Fire,}
\keyword{Soil,}
\keyword{Temperature}
