% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepthForTemp.R
\name{DepthForTemp}
\alias{DepthForTemp}
\title{Depth for Temperature}
\usage{
DepthForTemp(temp, inc = 0.5, res = 1, model)
}
\arguments{
\item{temp}{The temperature of interest.}

\item{inc}{The final increment (cm) used to calculate which depths to check (ie every 0.5 cm). Default is 0.5.}

\item{res}{The temporal resolution (in min) at which to calculate the temperatures. Default is one minute.}

\item{model}{The model object (a list) from the SheFire function containing the equations and data that comprise the model.}
}
\value{
This function returns the deepest depth (in cm) that reached the specified temperature. It also print a text string that contains the same information or reasons for a NA or NULL return.
}
\description{
Calculates the deepest depth (cm) that is reached by the specified temperature. When the exact depth is not tested, the function will choose the shallower of the two bounding depths (example: if 60C actually reaches 5.4cm but the function is set up such that it checks 5cm and 5.5cm, it will return 5cm). Using a smaller depth increment (inc) will increase precision.
}
\details{
Example data from US Forest Service FBAT: https://www.fs.fed.us/adaptivemanagement/reports/fbat/2019_FBATReport_WalkerFire_10112019_Final.pdf
}
\examples{
inputFile <- read.csv(file.path(system.file("external", package="SheFire"), "WalkerPlot4NE.csv")) #read in example data
model_example <- SheFire(input = inputFile )  #create model with all default settings

Deepest30 <- DepthForTemp(temp = 30, inc = 0.1, model = model_example)  #default temporal resolution res = 1
}
\seealso{
\code{\link{SheFire}}
}
\author{
MaryKBrady
}
\keyword{Fire,}
\keyword{Soil,}
\keyword{Temperature}
