% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatingPeriod.R
\name{HeatingPeriod}
\alias{HeatingPeriod}
\title{Heating Period}
\usage{
HeatingPeriod(x, res = 1, model)
}
\arguments{
\item{x}{The soil depth in cm.}

\item{res}{The temporal resolution (in min) at which to calculate the temperatures. Default is one minute.}

\item{model}{The model object (a list) from the SheFire function containing the equations and data that comprise the model.}
}
\value{
The function returns the temperatures at the set time interval (res) for the specified soil depth (x) from the point when it begins to warm through the temperature peak. It also prints a text string summarizing some information about the heating.
}
\description{
Isolates the portion of time when the soil depth of interest is heating and returns the temperature values for that period.
}
\details{
Example data from US Forest Service Fire Behavior Assessment Team: https://www.fs.fed.us/adaptivemanagement/reports/fbat/2019_FBATReport_WalkerFire_10112019_Final.pdf
For discussion on reasons to isolate just the heating limb of the temperature curve see: McGranahan, D. A. (2020). An inconvenient truth about temperature-time data from thermocouples. Plant Ecology, 221(11), 1091-1104.
}
\examples{
inputFile <- read.csv(file.path(system.file("external", package="SheFire"), "WalkerPlot4NE.csv")) #read in example data
model_example <- SheFire(input = inputFile )  #create model with all default settings

heating_portion <- HeatingPeriod(x = 4, model = model_example)  #res defaults to 1 minute
}
\seealso{
\code{\link{SheFire}}, \code{\link{CoolingPeriod}}
}
\author{
MaryKBrady
}
\keyword{Fire,}
\keyword{Soil,}
\keyword{Temperature}
