% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shefire.R
\name{shefire}
\alias{shefire}
\title{SheFire Model}
\usage{
shefire(
  input,
  sensor.depths = c(5, 10, 15),
  cutOff = 1440,
  override.clip = F,
  moving.window = F,
  window.size = 3,
  regression = T,
  res = 1,
  corr.threshold = 0.8,
  time.buffer = 30,
  print.plots.table = F,
  save.plots.table = F,
  save.name = "SheFire",
  save.directory = getwd()
)
}
\arguments{
\item{input}{Data.frame of data organized with columns: Date.Time, TimeCounter, Temp_S, Temp_M, Temp_D (columns must be in this order) where S is the shallow sensor, M is the middle sensor, and D is the deep sensor. Rows: each time point. TimeCounter must start at 1*TimeStep (data logging rate), not 0.}

\item{sensor.depths}{List of the temperature sensor depths in centimeters.}

\item{cutOff}{The time (min) after the shallow sensor's maximum temperature that the code will cut off the data set for fitting. It will use this cut off time unless the temps rise again (ie diurnal heating) or the data set ends first in which cases, those will be set as the end points. The default is 24 hours (1440 min) but look at your data and do what makes sense (Some hot, smoldering fires will be cooling for more than 24 hrs).}

\item{override.clip}{This parameter will prevent the function from clipping the data set. Default is False. Only change this if you have manually clipped the input data to exactly how you want it analyzed.}

\item{moving.window}{This option smooths the data when locating the starting point for heating in the shallow sensor. This can be useful in low temperature fires in particular, when the fire heating may not be faster than diurnal heating on a point to point basis. The default is False.}

\item{window.size}{The size (in time points) of the window that is used in smoothing with moving.window. Default is 3.}

\item{regression}{This will have the function calculate the parameter regressions which can be used to extrapolate or interpolate to other soil depths. Default is True. Set to False if you are only interested in the equations for the sensor depths.}

\item{res}{Only needed if you are not calculating the regressions. This sets the temporal resolution (in min) for the output temperatures from the equations fit to the sensor data. The temp values (at every time point based on the temporal resolution) are returned in the return list at the end of the function (ie temperature at every minute, or every half minute, or every 5 minutes, etc). Default is 1 minute.}

\item{corr.threshold}{Threshold for accepted level of correlation between the input data and the fitted BFD equations as well as between the calculated regression equations and the BFD parameters from the three sensor depths. Correlations less than the threshold will cause the function to throw an error message. Default is 0.8.}

\item{time.buffer}{The time (min) added before the temperature begins to rise for the shallow sensor. If you do need extrapolations to depths shallower than your shallow sensor or if you are not calculating the regressions, you can set time.buffer to 0. The buffer is needed to be able to extrapolate to shallower depths with the parameter regressions. You want to keep the time.buffer as short as possible because longer buffers can worsen the fits for equations and regressions. The default is set to 30 minutes.}

\item{print.plots.table}{Print the standard plots of data (plotted at the temporal resolution of the input data, not "res") and equations as well as the two summary tables containing equation details and fit information. Default is False.}

\item{save.plots.table}{Save the standard plots of data (plotted at the temporal resolution of the input data, not "res") and equations as jpeg images and the summary tables containing equation details and fit information as CSV files. Default is False.}

\item{save.name}{The name that will be used for saving the plots and tables. Default is "SheFire".}

\item{save.directory}{File path for where to save the plots and tables, if different from current working directory.}
}
\value{
If not calculating the regressions, the function returns a list of the equation values at time intervals set in res, the time resolution (res), the summary data.frame containing the equation parameters, and information about the time at the beginning and end of the clipped data set used to calculate those equations. If calculating the regressions, the function returns a list of all the equations and data necessary calculate temperature over time for a range of soil depths and run the application functions in this package. The list contains: BFDEquation - a function for calculating the temp over time given the BFD parameters; MaxTemp.reg, TimeAtMax.reg, Shape.reg, InitTemp.reg - functions to calculate the BFD parameters for a given soil depth; MaxTemp.coeffs, TimeAtMax.coeffs, Shape.coeffs, InitTemp.coeffs - the coefficients calculated for their respective parameter functions; InitTemp.byDepth - additional parameter needed for InitTemp.reg function, list of InitTemps calculated for the sensor depths; sensor.depths - additional parameter needed for InitTemp.reg function, list of sensor depths; Shallowest - shallowest depth (in cm) that it is mathematically possible for the model to calculate (note: shallowest depth for which the model is reasonable may be slightly deeper, see notes elsewhere in package about checking shallow depth for reasonability); FullTime - time in minutes that the model covers, StartTime - time at the beginning of the model time range, EndTime - time at the end of the model time range.
}
\description{
This function takes data from three co-located soil temperature sensors at different depths during a fire and fits an equation that closely matches the typical rise and fall of temperatures in soils heated by wildland fires. The equation originated in the building fire literature and is called the BFD equation (Barnett 2002). With parameter regression set to T, the model will also calculate regression equations for each of the BFD equation parameters. Those regressions enable temperature over time during the fire to be modeled at unmeasured soil depths.
}
\details{
Example data from US Forest Service Fire Behavior Assessment Team: https://www.fs.fed.us/adaptivemanagement/reports/fbat/2019_FBATReport_WalkerFire_10112019_Final.pdf
Papers Cited in this function: Barnett, C.R., 2002. BFD curve: a new empirical model for fire compartment temperatures. Fire Safety Journal, 37(5), pp.437-463.
}
\examples{
inputFile <- read.csv(file.path(system.file("external", package="SheFire"), "WalkerPlot4NE.csv")) #read in example data
model_example <- shefire(input = inputFile )  #create model with all default settings
}
\author{
MaryKBrady, based on work by Matthew B Dickinson
}
\keyword{Fire,}
\keyword{Soil,}
\keyword{Temperature}
