% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummDepthRange.R
\name{SummDepthRange}
\alias{SummDepthRange}
\title{Summarize a Depth Range}
\usage{
SummDepthRange(x, inc = 0.5, t = NULL, res = 1, model)
}
\arguments{
\item{x}{The soil depth in cm. Either one number or a vector of two numbers.}

\item{inc}{The increment (in cm) used to calculate which depths within the depth range to calculate.}

\item{t}{The time range, in min, (vector of two numbers or a single number if starting from the beginning) over which to calculate. Default is t = NULL, the full model time range.}

\item{res}{The temporal resolution (in min) at which to calculate the temperatures. Default is one minute.}

\item{model}{The model object (a list) from the SheFire function containing the equations and data that comprise the model.}
}
\value{
The function returns a data.frame of the mean, standard deviation, median, and maximum temperature for the depth range at each time point.
}
\description{
Calculates the mean temp, standard deviation, median temp, and max temp at each time point for a range of soil depths. The user can enter the shallow and deep bounds for the depth range or enter a single number and the range will default to that depth up to the shallowest depth calculable. NOTE: the deepest depth calculated will be the depth closest to the user defined max depth without exceeding it that can be reached by incrementing from the shallowest depth (example: if your shallow depth is 2.3cm and the increment is 0.5cm, the closest to a max depth of 4 would be 3.8cm). Once you have the values calculated at each time point, you can then run further statistics or functions on those temperatures.
}
\details{
Example data from US Forest Service Fire Behavior Assessment Team: https://www.fs.fed.us/adaptivemanagement/reports/fbat/2019_FBATReport_WalkerFire_10112019_Final.pdf
}
\examples{
inputFile <- read.csv(file.path(system.file("external", package="SheFire"), "WalkerPlot4NE.csv")) #read in example data
model_example <- SheFire(input = inputFile )  #create model with all default settings

Depth5to10cm <- SummDepthRange(x = c(5, 10), inc = 1, model = model_example) #t defaults to FullTime - full time range of model, res defaults to 1

#plot mean +/- stand dev over time
Depth5to10cm$time <- Depth5to10cm$time - 0.00001 #just to make it look a bit cleaner (model time starts at 0.00001, not 0, for math reasons)
ggplot(Depth5to10cm, (aes(x = time))) +   
 geom_line(aes(y = mean), color = "blue", size = 1) +
 geom_line(aes(y = mean - stnddev), color = "red", linetype = "dashed") +
 geom_line(aes(y = mean + stnddev), color = "red", linetype = "dashed") +
 labs(x = "Time (min)", y = "Temp (C)", title = "Mean temp for depths 5-10cm +/- standard deviation")
}
\seealso{
\code{\link{SheFire}}
}
\author{
MaryKBrady
}
\keyword{Fire,}
\keyword{Soil,}
\keyword{Temperature}
