% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalPercent.R
\name{SurvivalPercent}
\alias{SurvivalPercent}
\title{Survival Percent}
\usage{
SurvivalPercent(x, deltaS, deltaH, res = 1, temp = 48, model)
}
\arguments{
\item{x}{The soil depth in cm.}

\item{deltaS}{The change in entropy (units of J*mol-1*K-1) for the organism/tissue/etc of interest.}

\item{deltaH}{The change in enthalpy (units of J*mol-1) for the organism/tissue/etc of interest.}

\item{res}{The temporal resolution (in min) at which to calculate the temperatures. Default is one minute.}

\item{temp}{The threshold temperature when survival will begin to be impacted. Default is 48 based on Miesel et al (2020) and personal communication with Matthew Dickinson. See full works cited in package description.}

\item{model}{The model object (a list) from the SheFire function containing the equations and data that comprise the model.}
}
\value{
This function returns a list containing the final survival (1 - 100 percent, 0 - no survival) and a data.frame of survival through time (each row is a time point). It also prints a text string with survival information.
}
\description{
Models survival of the specified biological organism/tissue/protein/etc at the specified soil depth over time. The function requires two thermal tolerance parameters, deltaS and deltaH, specific to the organism/tissue/etc of interest. These parameters are specific to the kind of impact the user is looking at because different biological systems (seeds, roots, microorganisms, etc) have varying heat tolerances.
}
\details{
Example data from US Forest Service FBAT: https://www.fs.fed.us/adaptivemanagement/reports/fbat/2019_FBATReport_WalkerFire_10112019_Final.pdf
Other papers cited in this function: DICKINSON, M. B., & Johnson, E. A. (2004). Temperature-dependent rate models of vascular cambium cell mortality. Canadian Journal of Forest Research, 34(3), 546-559. ; MARTIN, R. E., Cushwa, C. T., & Miller, R. L. (1969). Fire as a physical factor in wildland management. In In: Proceedings Annual [9th] Tall Timbers Fire Ecology Conference. Tallahassee, FL. Tall Timbers Research, Inc., Tallahassee, FL. 271-288. (pp. 271-288). ; MIESEL, J. R., Dickinson, M. B., Kolka, R. K., Kern, C. C., Donner, D. M., Quigley, K. M., & Bushman, M. M. (2020). Manipulating soil heating patterns to optimize barrens restoration. FireScience.gov. ; ROSENGERG, B., Kemeny, G., Switzer, R. C., & Hamilton, T. C. (1971). Quantitative evidence for protein denaturation as the cause of thermal death. Nature, 232(5311), 471-473.
}
\examples{
inputFile <- read.csv(file.path(system.file("external", package="SheFire"), "WalkerPlot4NE.csv")) #read in example data
model_example <- SheFire(input = inputFile )  #create model with all default settings

CassiaSeeds2cm <- SurvivalPercent(x = 2, deltaS = 112, deltaH = 147741, model = model_example) #default temporal resolution and temp. Cassia Seed survival, for parameters see Martin et al (1969)
Proteins1.79cm <- SurvivalPercent(x = 1.79, deltaS = 482, deltaH = 248410, res = 0.5, model = model_example) #default temp. Protein thermal denaturation parameters, See Dickinson and Johnson (2004) and Rosenberg et al (1971)

#plot survival
PlotSurvival <- data.frame(Proteins1.79cm$table.Time, Proteins1.79cm$table.Survival)
colnames(PlotSurvival) <- c("Time", "Survival")
ggplot(PlotSurvival, aes (x = Time, y = Survival)) +
  geom_line() +
  ylim(0,1) +
  labs(x = "Time (min)", y = "Survival", title = "Protein Survival Over Time")
}
\seealso{
\code{\link{SheFire}}, \code{\link{SurvivalDepth}}
}
\author{
MaryKBrady, based on work by Matthew Dickinson
}
\keyword{Fire,}
\keyword{Soil,}
\keyword{Temperature}
