% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempIntervalsSet.R
\name{TempIntervalsSet}
\alias{TempIntervalsSet}
\title{Temperature Intervals Set}
\usage{
TempIntervalsSet(x, bins = c(0, 10, 20, 30, 40, 50, 60), res = 1, model)
}
\arguments{
\item{x}{The soil depth in cm. Either a single number or a vector of numbers.}

\item{bins}{A vector of the lower bounds of all the temperature bins.}

\item{res}{The temporal resolution (in min) at which to calculate the temperatures. Default is one minute.}

\item{model}{The model object (a list) from the SheFire function containing the equations and data that comprise the model.}
}
\value{
The function returns a data.frame with bin boundary temps and times in each bin for each depth.
}
\description{
Calculates time spent in different temp ranges (bins) for a specified soil depth(s). The user sets the lower bounds of each bin (bins don't have to be equal sizes). The final bin includes all temps over the last input temp (no upper bound to temps included in last bin).
}
\details{
Example data from US Forest Service FBAT: https://www.fs.fed.us/adaptivemanagement/reports/fbat/2019_FBATReport_WalkerFire_10112019_Final.pdf
}
\examples{
inputFile <- read.csv(file.path(system.file("external", package="SheFire"), "WalkerPlot4NE.csv")) #read in example data
model_example <- SheFire(input = inputFile )  #create model with all default settings

SetBins2cm <- TempIntervalsSet(x = 2, bins = c(10,15,20,25,30,35,40), model = model_example) #default res is 1 min

#example to compare diff depths with same temp bins - reformat table output from TempIntervalset function to plot
CompareDepths <- TempIntervalsSet(x = c(2,3,4,5), bins = seq(from = 10, to = 45, by = 5), model = model_example)
CompareDepths #view table
##to plot
timevalues <- as.numeric(c(CompareDepths[,3], CompareDepths[,4], CompareDepths[,5], CompareDepths[,6]))
type <- c(rep("Depth2cm", nrow(CompareDepths)), rep("Depth3cm", nrow(CompareDepths)), rep("Depth4cm", nrow(CompareDepths)), rep("Depth5cm", nrow(CompareDepths)))
binnames <- rep(c("10-14.9","15-19.9","20-24.9","25-29.9","30-34.9","35-39.9","40-44.9","45+"), 4)
TableToPlot <- data.frame(binnames, timevalues, type)
ggplot(TableToPlot, aes(binnames, timevalues)) +
  geom_col(position = "dodge", aes(fill = type)) +
  labs(x = "Temp bins", y = "Time (min)", title = "Time in temp ranges for different soil depths")
}
\seealso{
\code{\link{SheFire}}, \code{\link{TempIntervals}}
}
\author{
MaryKBrady
}
\keyword{Fire,}
\keyword{Soil,}
\keyword{Temperature}
