% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_temp_ranges.R
\name{time_temp_ranges}
\alias{time_temp_ranges}
\title{Time in Temperature Ranges}
\usage{
time_temp_ranges(x, bin = 5, res = 1, model)
}
\arguments{
\item{x}{The soil depth in cm.}

\item{bin}{The size, in degrees, of each temperature bin (ie bins of 5 degrees: 15-19.9, 20-24.9). Defaults is 5.}

\item{res}{The temporal resolution (in min) at which to calculate the temperatures. Default is one minute.}

\item{model}{The model object (a list) from the shefire function containing the equations and data that comprise the model.}
}
\value{
This function returns a data.frame of the bin temperature boundaries and the time (min) spent in each temp range.
}
\description{
Calculates total time spent in different temperature ranges (bins) for a specified soil depth. The bin boundary temperatures are set by the function depending on the temps reached at that depth and the bin size set by the user.
}
\details{
Example data from US Forest Service Fire Behavior Assessment Team: https://www.fs.fed.us/adaptivemanagement/reports/fbat/2019_FBATReport_WalkerFire_10112019_Final.pdf
}
\examples{
inputFile <- read.csv(file.path(system.file("external", package="SheFire"), "WalkerPlot4NE.csv")) #read in example data
model_example <- shefire(input = inputFile )  #create model with all default settings

binned5cm <- time_temp_ranges(x = 5, bin = 5, model = model_example) #res defaults to 1 minute
binned5cm.res <- time_temp_ranges(x = 5, bin = 5, res = 0.5, model = model_example) #changed temporal resolution to .5 min
}
\seealso{
\code{\link{shefire}}, \code{\link{set_temp_ranges}}
}
\author{
MaryKBrady
}
\keyword{Fire,}
\keyword{Soil,}
\keyword{Temperature}
