% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeAboveThreshold.R
\name{TimeAboveThreshold}
\alias{TimeAboveThreshold}
\title{Time Above Threshold}
\usage{
TimeAboveThreshold(x, temp, t = NULL, res = 1, model)
}
\arguments{
\item{x}{The soil depth in cm.}

\item{temp}{The temperature threshold.}

\item{t}{The time (min) for when to cut off the calculation (use a single number) or the time range (min) (use a list of two numbers) across which to calculate, defaults to the full model time period.}

\item{res}{The temporal resolution (in min) at which to calculate the temperatures. Default is one minute.}

\item{model}{The model object (a list) from the SheFire function containing the equations and data that comprise the model.}
}
\value{
Function returns the length of time in minutes that the soil depth spent at or above the threshold. The function also prints a text string with the same information.
}
\description{
Calculates the time (min) a specified soil depth spends at or above a specified temperature threshold.
}
\details{
Example data from US Forest Service FBAT: https://www.fs.fed.us/adaptivemanagement/reports/fbat/2019_FBATReport_WalkerFire_10112019_Final.pdf
}
\examples{
inputFile <- read.csv(file.path(system.file("external", package="SheFire"), "WalkerPlot4NE.csv")) #read in example data
model_example <- SheFire(input = inputFile )  #create model with all default settings

TimeAbove30 <- TimeAboveThreshold(x = 5, temp = 30, res = 1, model = model_example) #t defaults to FullTime - full time range of model
TimeAbove20 <- TimeAboveThreshold(x = 8, temp = 17, t = 10:1000, res = 0.5, model = model_example) #subset the time period used

#example with multiple depths
depths <- c(4, 6, 8)
Above25multiple <- sapply(X=depths, FUN = TimeAboveThreshold, temp = 25, t = 1100, model = model_example) #shortened time range
}
\seealso{
\code{\link{SheFire}}
}
\author{
MaryKBrady
}
\keyword{Fire,}
\keyword{Soil,}
\keyword{Temperature}
